# lib/api-zod/src/generated/types/researchFeedBusiness.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Local business discovery engine API
 * OpenAPI spec version: 0.2.0
 */

export interface ResearchFeedBusiness {
  id: number;
  name: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  country?: string | null;
  categorySlug: string;
  /** @nullable */
  targetMarket?: string | null;
  /** @nullable */
  avatarType?: string | null;
  /** @nullable */
  engineType?: string | null;
  /** @nullable */
  targetType?: string | null;
  /** @nullable */
  targetCluster?: string | null;
  /** @nullable */
  economicValueScore?: number | null;
  /** @nullable */
  strategicValueScore?: number | null;
  /** @nullable */
  offerFitScore?: number | null;
  /** @nullable */
  relationshipPathScore?: number | null;
  /** @nullable */
  actionabilityScore?: number | null;
  warmPathExists?: boolean;
  readyForRelationship?: boolean;
  readyForInstitutionalPitch?: boolean;
  prestigeWatchlist?: boolean;
  cultivationRequired?: boolean;
  /** @nullable */
  relevanceScore?: number | null;
  /** @nullable */
  contactabilityScore?: number | null;
  /** @nullable */
  confidenceScore?: number | null;
  /** @nullable */
  freshnessScore?: number | null;
  /** @nullable */
  priorityScore?: number | null;
  /** @nullable */
  researchScore?: number | null;
  readyForOutreach: boolean;
  reviewRequired: boolean;
  /** @nullable */
  reviewReason?: string | null;
  /** @nullable */
  recommendedNextStep?: string | null;
  /** @nullable */
  topGap?: string | null;
}